Decompose server.ts
